+++ mitmproxy/addons/serverplayback.py
@@ -30,7 +30,16 @@ class ServerPlayback:
             "server_replay_kill_extra",
             bool,
             False,
-            "Kill extra requests during replay (for which no replayable response was found).",
+            "Kill extra requests during replay (for which no replayable response was found)."
+            "[Deprecated, prefer to use server_replay_extra='kill']",
+        )
+        loader.add_option(
+            "server_replay_extra",
+            str,
+            "forward",
+            "Behaviour for extra requests during replay for which no replayable response was found. "
+            "Setting a numeric string value will return an empty HTTP response with the respective status code.",
+            choices=["forward", "kill", "204", "400", "404", "500"],
         )
         loader.add_option(
             "server_replay_reuse",
@@ -230,6 +239,11 @@ class ServerPlayback:
             return None
 
     def configure(self, updated):
+        if ctx.options.server_replay_kill_extra:
+            logger.warning(
+                "server_replay_kill_extra has been deprecated, "
+                "please update your config to use server_replay_extra='kill'."
+            )
         if ctx.options.server_replay_nopop:  # pragma: no cover
             logger.error(
                 "server_replay_nopop has been renamed to server_replay_reuse, please update your config."
@@ -252,10 +266,21 @@ class ServerPlayback:
                     response.refresh()
                 f.response = response
                 f.is_replay = "response"
-            elif ctx.options.server_replay_kill_extra:
+            elif (
+                ctx.options.server_replay_kill_extra
+                or ctx.options.server_replay_extra == "kill"
+            ):
                 logging.warning(
                     "server_playback: killed non-replay request {}".format(
                         f.request.url
                     )
                 )
                 f.kill()
+            elif ctx.options.server_replay_extra != "forward":
+                logging.warning(
+                    "server_playback: returned {} non-replay request {}".format(
+                        ctx.options.server_replay_extra, f.request.url
+                    )
+                )
+                f.response = http.Response.make(int(ctx.options.server_replay_extra))
+                f.is_replay = "response"